Move path-related fixtures into their own module
neon_fixtures.py has grown into quite a beast.
